refactor: 레거시 코드 리팩토링
