+++ mvc/src/main/java/nextstep/mvc/controller/HandlerExecutionHandlerAdapter.java
@@ -1,23 +1,20 @@
-package nextstep.mvc.controller.asis;
+package nextstep.mvc.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
-import nextstep.mvc.controller.asis.Controller;
-import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 
-public class ManualHandlerAdapter implements HandlerAdapter {
+public class HandlerExecutionHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(final Object handler) {
-        return handler instanceof Controller;
+        return handler instanceof HandlerExecution;
     }
 
     @Override
     public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response,
                                final Object handler) throws Exception {
-        final String viewPath = ((Controller) handler).execute(request,response);
-        return new ModelAndView(new JspView(viewPath));
+        return ((HandlerExecution) handler).handle(request, response);
     }
 }
