+++ mvc/src/test/java/samples/ManualTestHandlerAdapter.java
@@ -1,20 +1,22 @@
-package nextstep.mvc.controller.tobe;
+package samples;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
+import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
 
-public class HandlerExecutionHandlerAdapter implements HandlerAdapter {
+public class ManualTestHandlerAdapter implements HandlerAdapter {
 
     @Override
     public boolean supports(final Object handler) {
-        return handler instanceof HandlerExecution;
+        return handler instanceof Controller;
     }
 
     @Override
     public ModelAndView handle(final HttpServletRequest request, final HttpServletResponse response,
                                final Object handler) throws Exception {
-        return ((HandlerExecution) handler).handle(request, response);
+        final String viewPath = ((Controller) handler).execute(request,response);
+        return new ModelAndView(new JspView(viewPath));
     }
 }
